덱 카드 초기값 설정을 위해 Suit, Number를 따로 밖으로 꺼냈습니다.
